+++ actionpack/test/controller/render_test.rb
@@ -290,6 +290,24 @@ class TestController < ActionController::Base
     end
   end
 
+  def cache_control_default_header_with_extras_partially_overridden_by_expires_in
+    response.headers["Cache-Control"] = "max-age=120, public, s-maxage=60, proxy-revalidate"
+    expires_in 300.seconds, public: true
+    render action: "hello_world"
+  end
+
+  def cache_control_no_store_overridden_by_expires_in
+    response.headers["Cache-Control"] = "no-store"
+    expires_in 60.seconds, public: true
+    render action: "hello_world"
+  end
+
+  def cache_control_no_store_overridden_by_expires_now
+    response.headers["Cache-Control"] = "no-store"
+    expires_now
+    render action: "hello_world"
+  end
+
   private
     def set_variable_for_layout
       @variable_for_layout = nil
@@ -462,6 +480,21 @@ class ExpiresInRenderTest < ActionController::TestCase
       get :conditional_hello_with_expires_in
       assert_equal Time.now.httpdate, @response.headers["Date"]
     end
+  end
+
+  def test_cache_control_default_header_with_extras_partially_overridden_by_expires_in
+    get :cache_control_default_header_with_extras_partially_overridden_by_expires_in
+    assert_equal "max-age=300, public, s-maxage=60, proxy-revalidate", @response.headers["Cache-Control"]
+  end
+
+  def test_cache_control_no_store_overridden_by_expires_in
+    get :cache_control_no_store_overridden_by_expires_in
+    assert_equal "max-age=60, public", @response.headers["Cache-Control"]
+  end
+
+  def test_cache_control_no_store_overridden_by_expires_now
+    get :cache_control_no_store_overridden_by_expires_now
+    assert_equal "no-cache", @response.headers["Cache-Control"]
   end
 end
 
@@ -987,5 +1020,109 @@ class HttpCacheForeverTest < ActionController::TestCase
     @request.if_none_match = @response.etag
     get :cache_me_forever
     assert_response :not_modified
+  end
+end
+
+class HttpCacheNoStoreTest < ActionController::TestCase
+  class HttpCacheNoStoreController < ActionController::Base
+    def standalone_no_store_call
+      no_store
+      render plain: "hello world"
+    end
+
+    before_action(only: :no_store_overridden_by_expires_in) { no_store }
+    def no_store_overridden_by_expires_in
+      expires_in 30.seconds
+      render plain: "hello world"
+    end
+
+    before_action(only: :expires_in_overridden_by_no_store) { expires_in 30.seconds }
+    def expires_in_overridden_by_no_store
+      no_store
+      render plain: "hello world"
+    end
+
+    before_action(only: :no_store_overridden_by_fresh_when) { no_store }
+    def no_store_overridden_by_fresh_when
+      fresh_when(etag: "123abc")
+      render plain: "hello world"
+    end
+
+    before_action(only: :fresh_when_overridden_by_no_store) { fresh_when etag: "abc123" }
+    def fresh_when_overridden_by_no_store
+      no_store
+      render plain: "hello world"
+    end
+
+    before_action(only: :expires_now_overridden_by_no_store) { expires_now }
+    def expires_now_overridden_by_no_store
+      no_store
+      render plain: "hello world"
+    end
+
+    before_action(only: :no_store_overridden_by_expires_now) { no_store }
+    def no_store_overridden_by_expires_now
+      expires_now
+      render plain: "hello world"
+    end
+
+    def cache_control_no_cache_overridden_by_no_store
+      response.headers["Cache-Control"] = "no-cache"
+      no_store
+      render plain: "hello world"
+    end
+
+    def cache_control_public_with_max_age_overridden_by_no_store
+      response.headers["Cache-Control"] = "public, max-age=604800"
+      no_store
+      render plain: "hello world"
+    end
+  end
+
+  tests HttpCacheNoStoreController
+
+  def test_standalone_no_store_call
+    get :standalone_no_store_call
+    assert_equal "no-store", @response.headers["Cache-Control"]
+  end
+
+  def test_no_store_overridden_by_expires_in
+    get :no_store_overridden_by_expires_in
+    assert_equal "max-age=30, private", @response.headers["Cache-Control"]
+  end
+
+  def test_expires_in_overridden_by_no_store
+    get :expires_in_overridden_by_no_store
+    assert_equal "no-store", @response.headers["Cache-Control"]
+  end
+
+  def test_no_store_overridden_by_fresh_when
+    get :no_store_overridden_by_fresh_when
+    assert_equal "max-age=0, private, must-revalidate", @response.headers["Cache-Control"]
+  end
+
+  def test_fresh_when_overridden_by_no_store
+    get :fresh_when_overridden_by_no_store
+    assert_equal "no-store", @response.headers["Cache-Control"]
+  end
+
+  def test_expires_now_overridden_by_no_store
+    get :expires_now_overridden_by_no_store
+    assert_equal "no-store", @response.headers["Cache-Control"]
+  end
+
+  def test_no_store_overridden_by_expires_now
+    get :no_store_overridden_by_expires_now
+    assert_equal "no-cache", @response.headers["Cache-Control"]
+  end
+
+  def test_cache_control_no_cache_header_can_be_overridden_by_no_store
+    get :cache_control_no_cache_overridden_by_no_store
+    assert_equal "no-store", @response.headers["Cache-Control"]
+  end
+
+  def test_cache_control_public_with_expiration_header_can_be_overridden_by_no_store
+    get :cache_control_public_with_max_age_overridden_by_no_store
+    assert_equal "no-store", @response.headers["Cache-Control"]
   end
 end